pyproject.toml
[tool.poetry]
name = "00-airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["ssu-uky <id_suhyun@naver.com>"]
readme = "README.md"
packages = [{include = "00_airbnb_clone_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
